requirements.txt
pandas
pytest
streamlit
sentence-transformers
groq
python-dotenv
